Move bracket layer code to own module
